+++ app/widget/plots_placeholder_widget.py
